fix: Do not interpret '#' symbols in variable values as comments (use env)
Cronrunner used to pass variables to commands like this:

```shell
FOO=bar;echo "job command"
```

But this fails in cases like this one:

```crontab
MY_VAR=contains a '#' symbol

@daily echo $MY_VAR
```

Because the '#' symbol makes everything after it a comment, the command
is not even run:

```shell
MY_VAR=contains a '#' symbol;echo $MY_VAR
```

But with Cron, '#' is interpreted literally, and the whole value is
used.

With this fix, cronrunner does too. Variables are put in the environment
instead of the command. Cases like the one above work fine now.