refactor!: Rename `cr` executable to `crn`
This is something that's been bothering me from the moment I first
renamed the `cronrunner` executable to `cr`.

Two-letter executable names should be reserved for "core" command.
Cronrunner is not something you'd use as much as `cd` or `ls`, typing a
third letter is _fine_.

This technically constitutes a breaking change, but I won't bother with
a major version bump, as the executable can be easily renamed or aliased
if needed.